+++ src/thunks/fetchStarIds.js
